청소 명령어 유지보수: 에러 메시지 추가

+++ functions/manage.py
@@ -27,18 +27,28 @@ class manage(commands.Cog):
 
     @commands.command()
     async def 청소(self, ctx, count: int):
-        if ctx.author.guild_permissions.administrator:
-            await asyncio.sleep(2)
-            await ctx.channel.purge(limit=count + 1)
-            embed=nextcord.Embed(title="청소 완료!", color=0xffffff)
-            embed.add_field(name="삭제한 메시지의 수:", value=f"{count}", inline=False)
-            erasemsg = await ctx.send(embed=embed)
-            await asyncio.sleep(3)
-            await erasemsg.delete()
-        else:
+        try:
+            if ctx.author.guild_permissions.administrator:
+                if count:
+                    await asyncio.sleep(2)
+                    await ctx.channel.purge(limit=count + 1)
+                    embed=nextcord.Embed(title="청소 완료!", color=0xffffff)
+                    embed.add_field(name="삭제한 메시지의 수:", value=f"{count}", inline=False)
+                    erasemsg = await ctx.send(embed=embed)
+                    await asyncio.sleep(3)
+                    await erasemsg.delete()
+                else:
+                    embed=nextcord.Embed(title="오류 발생!", color=0xff0000)
+                    embed.add_field(name="값 오류", value="올바른 자연수 값을 입력해주세요.")
+            else:
+                embed=nextcord.Embed(title="오류 발생!", color=0xff0000)
+                embed.add_field(name="권한 오류", value="권한 확인 후 다시 시도해주세요.", inline=False)
+                await ctx.send(embed=embed)
+        except:
             embed=nextcord.Embed(title="오류 발생!", color=0xff0000)
-            embed.add_field(name="메시지 관리 권한이 필요합니다.", value="권한 확인 후 다시 시도해주세요.", inline=False)
+            embed.add_field(name="권한 오류", value="권한 확인 후 다시 시도해주세요.", inline=False)
             await ctx.send(embed=embed)
+
 
 
     @commands.command()
@@ -59,7 +69,7 @@ class manage(commands.Cog):
                 )
         else:
             await ctx.send(
-                f'{ctx.author.mention} 관리자 권한이 필요합니다. 권한 확인 후 다시 실행해주세요.')
+                f'{ctx.author.mention} 멤버 추방하기 권한이 필요합니다. 권한 확인 후 다시 실행해주세요.')
 
 
     @commands.command()
